added memory overflow check in flowlines stepping, changed data type in magnitude fct from int to float, added random shape profil option, added ploting fct for glacier shape

+++ combine/core/arithmetics.py
@@ -60,7 +60,7 @@ def to_numpy_array(val):
 def magnitude(x):
     x = np.where(x != 0,
                  x, 1e-100)
-    return np.floor(np.log10(np.abs(x))).astype(int)
+    return np.floor(np.log10(np.abs(x))).astype(float)
 
 
 class para_width_from_thick(Function):
